requirements.txt
aiogram==3.5
fastapi==0.115
uvicorn==0.32
python-dotenv==1.0
